requirements.txt
alabaster==0.7.12
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.12.9
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
awesome-slugify==1.6.5
Babel==2.10.3
backcall==0.2.0
black==22.6.0
boto3==1.24.78
botocore==1.27.78
certifi==2022.9.14
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
Collectfast==2.2.0
colorama==0.4.5
coverage==6.4.1
crispy-bootstrap5==0.6
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.6
Django==3.2.14
django-allauth==0.51.0
django-anymail==8.6
django-bootstrap-icons==0.8.1
django-ckeditor==6.5.1
django-coverage-plugin==2.0.3
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-djeddit==0.3.4
django-environ==0.9.0
django-etc==1.4.0
django-extensions==3.1.5
django-hitcount==1.3.5
django-ipware==4.0.2
django-js-asset==2.0.0
django-meta==2.1.0
django-model-utils==4.2.0
django-mptt==0.14.0
django-redis==5.2.0
django-storages==1.13.1
django-stubs==1.12.0
django-stubs-ext==0.5.0
docutils==0.18.1
executing==1.0.0
factory-boy==3.2.1
Faker==14.2.0
filelock==3.8.0
flake8==4.0.1
flake8-isort==4.1.1
gunicorn==20.1.0
hiredis==2.0.0
identify==2.5.5
idna==3.4
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.5.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.7.1
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.24.1
oauthlib==3.2.1
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.4.0
pylint==2.15.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytorch-pretrained-bert==0.6.2
pytz==2022.1
PyYAML==6.0
redis==4.3.3
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.0.1
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.5.0
termcolor==2.0.1
testfixtures==6.18.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
types-pytz==2022.2.1.0
types-PyYAML==6.0.11
typing_extensions==4.3.0
Unidecode==0.4.21
urllib3==1.26.12
virtualenv==20.16.5
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.8.1
